Simplify pbf callbacks
