+++ include/osm_store.h
@@ -1,10 +1,6 @@
 #ifndef _OSM_STORE_H
 #define _OSM_STORE_H
 
-// Protobuf
-#include "osmformat.pb.h"
-#include "vector_tile.pb.h"
-
 /**
 	OSM Store
 
@@ -159,28 +155,10 @@ public:
 	size_t size();
 };
 
-class PbfReaderOutput
-{
-public:
-	virtual void everyNode(NodeID id, LatpLon node) {};
-
-	// We are now processing a node
-	virtual void setNode(NodeID id, DenseNodes *dPtr, int kvStart, int kvEnd, LatpLon node, const std::map<std::string, std::string> &tags) {};
-
-	// We are now processing a way
-	virtual void setWay(Way *way, NodeVec *nodeVecPtr, bool inRelation, const std::map<std::string, std::string> &tags) {};
-
-	// We are now processing a relation
-	// (note that we store relations as ways with artificial IDs, and that
-	//  we use decrementing positive IDs to give a bit more space for way IDs)
-	virtual void setRelation(Relation *relation, WayVec *outerWayVecPtr, WayVec *innerWayVecPtr,
-		const std::map<std::string, std::string> &tags) {};
-};
-
 //
 // OSM store, containing all above.
 //
-class OSMStore : public PbfReaderOutput
+class OSMStore
 {
 
 public:
@@ -306,8 +284,6 @@ public:
 	Linestring nodeListLinestring(WayID wayId) const;
 
 	Linestring nodeListLinestring(const NodeVec &nodeVec) const;
-
-	virtual void everyNode(NodeID id, LatpLon node);
 
 private:
 	// helper
